Split docker executor code to its own file to reduce complexity
